chore(firstNonRepeat/consolidate): Tests and Problems together
One project for test and problem.
