+++ InterviewQuestions.sln
@@ -53,8 +53,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FirstNonRepeatingCharactersInStringTester", "FirstNonRepeatingCharactersInStringTester\FirstNonRepeatingCharactersInStringTester.csproj", "{1C1E38DA-7283-434C-A626-6AFB4DA697A7}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EggDropTrialQuestion", "EggDropTrialQuestion\EggDropTrialQuestion.csproj", "{C788EAE6-D9DD-473F-A551-62DFD2EE2689}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListMergeSort", "LinkedListMergeSort\LinkedListMergeSort.csproj", "{C6226CB8-76F1-476F-A13E-C0F4F44490FD}"
@@ -173,10 +171,6 @@ Global
 		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6B3B757F-DEA0-457C-91AB-3743AECC25F2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1C1E38DA-7283-434C-A626-6AFB4DA697A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C1E38DA-7283-434C-A626-6AFB4DA697A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C1E38DA-7283-434C-A626-6AFB4DA697A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C1E38DA-7283-434C-A626-6AFB4DA697A7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C788EAE6-D9DD-473F-A551-62DFD2EE2689}.Release|Any CPU.ActiveCfg = Release|Any CPU
