+++ NonRepeatingCharacter/FirstNonRepeatingCharacterInStringQuestion.csproj
@@ -33,6 +33,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="nunit.framework, Version=3.9.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
+      <HintPath>..\packages\NUnit.3.9.0\lib\net45\nunit.framework.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -43,14 +46,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FirstNonRepeatingCharacterInString.cs" />
+    <Compile Include="Problem\FirstNonRepeatingCharacterInString.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StringsLengths.cs" />
+    <Compile Include="Problem\StringsLengths.cs" />
+    <Compile Include="Tests\FirstNonRepeatingCharInStringTester.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="packages.config" />
     <None Include="README.md" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
