+++ NonRepeatingCharacter/packages.config
